// src/generated/types.ts
/**
 * Auto-generated TypeScript bindings for Tauri commands
 * Generated by tauri-typegen v0.5.0
 * Generated at: 2026-05-10T11:21:44.728292+00:00
 * Generator: none
 *
 * Do not edit manually - regenerate using: cargo tauri-typegen generate
 */



export interface DividendData {
  forward_dividend_rate?: number | null;
  forward_dividend_yield?: number | null;
  ex_dividend_date?: number | null;
  payout_ratio?: number | null;
  trailing_dividend_rate?: number | null;
  trailing_dividend_yield?: number | null;
}

export interface Info {
  header: HeaderData;
  summary: SummaryData;
  about: AboutData;
  dividends: DividendData;
}

export interface HeaderData {
  name?: string | null;
  kind?: string | null;
  symbol?: string | null;
  exchange?: string | null;
  currency?: string | null;
}

export interface TradeData {
  account_id: string;
  symbol: string;
  trade_type: string;
  date: string;
  quantity: number;
  price: number;
  costs: number;
}

export interface SearchResult {
  symbol: string;
  name: string;
  kind: string;
  exchange: string;
}

export interface Account {
  name: string;
  account_id: string;
}

export interface SummaryData {
  previous_close?: number | null;
  open?: number | null;
  day_low?: number | null;
  day_high?: number | null;
  beta?: number | null;
  volume?: number | null;
  average_volume?: number | null;
  bid?: number | null;
  ask?: number | null;
  bid_size?: number | null;
  ask_size?: number | null;
  market_cap?: number | null;
  fifty_two_week_low?: number | null;
  fifty_two_week_high?: number | null;
  trailing_price_to_earnings?: number | null;
  trailing_price_to_sales?: number | null;
}

export interface AboutData {
  description?: string | null;
  industry?: string | null;
  sector?: string | null;
  employees?: number | null;
  country?: string | null;
  fund_family?: string | null;
  fund_inception_date?: number | null;
  legal_type?: string | null;
}

export interface InfoParams {
  symbol: string;
  [key: string]: unknown;
}


export interface AddTradeParams {
  data: TradeData;
  [key: string]: unknown;
}


export interface SearchParams {
  query: string;
  [key: string]: unknown;
}


export interface AddAccountParams {
  name: string;
  [key: string]: unknown;
}


export interface RenameAccountParams {
  accountId: string;
  name: string;
  [key: string]: unknown;
}


export interface DeleteAccountParams {
  accountId: string;
  [key: string]: unknown;
}